refactor(economy): phase 3 slice 6 treasury cluster and town tile walk (#3939)
Consolidate ten treasury test_support modules under treasury_scenarios/ with
shared fixtures; extract forEachTownConnectedTileInProvince for town-bonus
GP/non-GP province loops. All 523 economy tests and CI gates pass.

Refs #3939
