+++ packages/colonizethis_economy/lib/src/economy/town_manufacturing_bonus.dart
@@ -9,6 +9,7 @@ import 'game_lookup_helpers.dart';
 import 'resource_extractor.dart';
 import 'sea_transport.dart';
 import 'tile_extraction_pipeline.dart';
+import 'town_connected_tile_walk.dart';
 
 /// Town manufacturing clustering bonus (Refs #3872).
 /// SPEC/game/extraction-and-improvements.md § Town manufacturing bonus.
@@ -161,41 +162,42 @@ void _accumulateGpProvinceExtraction({
       game.worldState.playerProspectedTiles[player.id] ?? const {};
   final capitalRegionId = player.capitalTile?.regionId;
 
-  for (final tileKey in connected) {
-    if (!townConnected.contains(tileKey)) continue;
-    final tileProvinceId = _fullProvinceIdFromTileKey(tileKey);
-    if (tileProvinceId != province.id) continue;
-
-    final contribution = computeTileExtractionContributionForPlayer(
-      game: game,
-      tileMapByRegion: tileMapByRegion,
-      player: player,
-      tileKey: tileKey,
-      connectedTileKeys: connected,
-      pathTransportCap: pathTransportCap,
-      connectedByRoadRule: roadRuleTiles,
-      portTileKeys: portTileKeys,
-      prospectedTileKeys: prospected,
-      capitalRegionId: capitalRegionId,
-      techCapForPlayer: (_) => extractionCapForUnlocked(player.techUnlocked),
-      techCapForPlayerAndResource: (_, resourceId) =>
-          extractionCapForResourceForUnlocked(player.techUnlocked, resourceId),
-      provincesByFullId: provincesByFullId,
-    );
-    if (contribution == null || contribution.units <= 0) continue;
-
-    if (contribution.isLandRelativeToCapital) {
-      final totals = landByProvince.putIfAbsent(province.id, () => {});
-      addUnits(totals, contribution.commodityId, contribution.units);
-    } else {
-      final byProvince = overseasByPlayerProvince.putIfAbsent(
-        player.id,
-        () => {},
+  forEachTownConnectedTileInProvince(
+    connectedTiles: connected,
+    townConnected: townConnected,
+    provinceId: province.id,
+    onTile: (tileKey) {
+      final contribution = computeTileExtractionContributionForPlayer(
+        game: game,
+        tileMapByRegion: tileMapByRegion,
+        player: player,
+        tileKey: tileKey,
+        connectedTileKeys: connected,
+        pathTransportCap: pathTransportCap,
+        connectedByRoadRule: roadRuleTiles,
+        portTileKeys: portTileKeys,
+        prospectedTileKeys: prospected,
+        capitalRegionId: capitalRegionId,
+        techCapForPlayer: (_) => extractionCapForUnlocked(player.techUnlocked),
+        techCapForPlayerAndResource: (_, resourceId) =>
+            extractionCapForResourceForUnlocked(player.techUnlocked, resourceId),
+        provincesByFullId: provincesByFullId,
       );
-      final totals = byProvince.putIfAbsent(province.id, () => {});
-      addUnits(totals, contribution.commodityId, contribution.units);
-    }
-  }
+      if (contribution == null || contribution.units <= 0) return;
+
+      if (contribution.isLandRelativeToCapital) {
+        final totals = landByProvince.putIfAbsent(province.id, () => {});
+        addUnits(totals, contribution.commodityId, contribution.units);
+      } else {
+        final byProvince = overseasByPlayerProvince.putIfAbsent(
+          player.id,
+          () => {},
+        );
+        final totals = byProvince.putIfAbsent(province.id, () => {});
+        addUnits(totals, contribution.commodityId, contribution.units);
+      }
+    },
+  );
 }
 
 void _accumulateNonGpProvinceExtraction({
@@ -213,31 +215,32 @@ void _accumulateNonGpProvinceExtraction({
   final capitalRegionId = _capitalRegionIdForOwner(game, ownerId);
   if (capitalProvinceId == null || capitalRegionId == null) return;
 
-  for (final tileKey in connectivity.connected) {
-    if (!townConnected.contains(tileKey)) continue;
-    final tileProvinceId = _fullProvinceIdFromTileKey(tileKey);
-    if (tileProvinceId != province.id) continue;
-
-    final contribution = computeTileYieldContribution(
-      game: game,
-      tileMapByRegion: tileMapByRegion,
-      tileKey: tileKey,
-      connectedTileKeys: connectivity.connected,
-      pathTransportCap: connectivity.pathTransportCap,
-      connectedByRoadRule: connectivity.connectedByRoadRule,
-      portTileKeys: portTileKeys,
-      capitalProvinceId: capitalProvinceId,
-      capitalRegionId: capitalRegionId,
-      logContext: 'town_bonus_non_gp',
-      provincesByFullId: provincesByFullId,
-      techCapForCommodity: (_) => defaultExtractionCap,
-      isCommodityExtractable: (tileKey, commodityId) =>
-          !kMineralResourceIds.contains(commodityId),
-    );
-    if (contribution == null || contribution.units <= 0) continue;
-    final totals = landByProvince.putIfAbsent(province.id, () => {});
-    addUnits(totals, contribution.commodityId, contribution.units);
-  }
+  forEachTownConnectedTileInProvince(
+    connectedTiles: connectivity.connected,
+    townConnected: townConnected,
+    provinceId: province.id,
+    onTile: (tileKey) {
+      final contribution = computeTileYieldContribution(
+        game: game,
+        tileMapByRegion: tileMapByRegion,
+        tileKey: tileKey,
+        connectedTileKeys: connectivity.connected,
+        pathTransportCap: connectivity.pathTransportCap,
+        connectedByRoadRule: connectivity.connectedByRoadRule,
+        portTileKeys: portTileKeys,
+        capitalProvinceId: capitalProvinceId,
+        capitalRegionId: capitalRegionId,
+        logContext: 'town_bonus_non_gp',
+        provincesByFullId: provincesByFullId,
+        techCapForCommodity: (_) => defaultExtractionCap,
+        isCommodityExtractable: (tileKey, commodityId) =>
+            !kMineralResourceIds.contains(commodityId),
+      );
+      if (contribution == null || contribution.units <= 0) return;
+      final totals = landByProvince.putIfAbsent(province.id, () => {});
+      addUnits(totals, contribution.commodityId, contribution.units);
+    },
+  );
 }
 
 void _applyOverseasCargoDeliveryToProvinces({
@@ -441,10 +444,4 @@ String? _capitalRegionIdForOwner(Game game, String ownerId) {
     if (tribe.id == ownerId) return tribe.capitalTile?.regionId;
   }
   return null;
-}
-
-String? _fullProvinceIdFromTileKey(String tileKey) {
-  final coords = parseTileKeyCoordinates(tileKey);
-  if (coords == null) return null;
-  return ProvinceId.full(coords.regionId, coords.provinceLocalId);
 }
