+++ packages/colonizethis_economy_test_support/lib/src/sellable_quantity_scenarios.dart
@@ -4,7 +4,7 @@ import 'package:colonizethis_economy/colonizethis_economy.dart';
 import 'package:colonizethis_models/colonizethis_models.dart';
 import 'package:colonizethis_test/test.dart';
 
-import 'treasury_bid_budget_test_support.dart';
+import 'treasury_scenarios/treasury_test_support.dart';
 
 /// One row in [offerCapByCommodityIdScenarios].
 typedef OfferCapByCommodityIdScenario = ({
