+++ packages/colonizethis_economy_test_support/lib/src/treasury_scenarios/treasury_scenarios.dart
@@ -1,0 +1,6 @@
+// Barrel for consolidated treasury scenario modules (Refs #3939 phase 3 slice 6).
+
+export 'treasury_available_scenarios.dart';
+export 'treasury_bid_spend_scenarios.dart';
+export 'treasury_pure_helper_scenarios.dart';
+export 'treasury_test_support.dart';
